+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 2.3.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Content listing: Allow to use collection query parameters from context.
+  Content listing: Include query parameters from request.
+  Content listing: Add "tile_class".
+  [agitator]
 
 
 2.3.2 (2019-12-05)